LayoutTests/imported/w3c/web-platform-tests/css/mediaqueries/forced-colors-expected.txt


FAIL Should be known: '(forced-colors)' assert_true: Can parse with JS expected true got false
FAIL Should be known: '(forced-colors: none)' assert_true: Can parse with JS expected true got false
FAIL Should be known: '(forced-colors: active)' assert_true: Can parse with JS expected true got false
FAIL Should be unknown: '(forced-colors: 0)' assert_true: Can parse with JS expected true got false
FAIL Should be unknown: '(forced-colors: no-preference)' assert_true: Can parse with JS expected true got false
FAIL Should be unknown: '(forced-colors: 10px)' assert_true: Can parse with JS expected true got false
FAIL Should be unknown: '(forced-colors: active 0)' assert_true: Can parse with JS expected true got false
FAIL Should be unknown: '(forced-colors: none active)' assert_true: Can parse with JS expected true got false
FAIL Should be unknown: '(forced-colors: active/none)' assert_true: Can parse with JS expected true got false
FAIL Check that none evaluates to false in the boolean context assert_equals: expected true but got false
